Some changes in double

+++ database/migrations/2016_12_11_212503_create_double_games_table.php
@@ -17,8 +17,8 @@ class CreateDoubleGamesTable extends Migration {
             $table->increments('id');
             $table->integer('num')->unsigned()->nullable();
             $table->integer('status')->unsigned();
-            $table->bigInteger('price');
-            $table->integer('am');
+            $table->float('price');
+            $table->float('am');
         });
     }
